Slightly refactor IsSupportedF8Pattern in gemm_rewriter.cc.
The overall logic is a bit more clear now. A type InstrPath is added, aliasing the type 'std::vector<std::pair<HloInstruction *, int>>' which was previously passed around. This type is now always ordered from operand to user instead of being reversed within IsSupportedF8Pattern. The comments for several functions were expanded. FindF8SubgraphRecursive now only returns an InstrPath instead of both returning one and taking one in.

The one functional change is that now we pattern match the following into an FP8 cublasLT custom call

  dot(x.astype(jnp.float16), y.astype(jnp.float16))

We previously matched a version of this that also included scaling into a custom call.

  dot(x.astype(jnp.float16) * x_scale, y.astype(jnp.float16) * y_scale)

Now the scales are optional even when the inputs are cast to FP16. There is no particular reason to cast inputs to FP16 if there are no input scales, but at least one internal model does this.

PiperOrigin-RevId: 636735933